- Changing model names to be a bit more clear - Renaming Grid* components to Table* - Starting to migrate the common code from GridTable.vue to dataTable.ts, so it can be reused in TreeTable.vue
