+++ api/src/diff/diff.service.ts
@@ -1,21 +1,21 @@
 ﻿import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";
 import { getChanges } from "src/diff/diff.utilities";
-import { TDiffValue } from "src/models/diff.entity";
+import { TreeChangeModel } from "src/models/diff.entity";
 
 @Injectable()
 export class DiffService {
 	getDiff() {
-		const records: Record<string, TDiffValue>[] = [];
+		const records: TreeChangeModel[] = [];
 		for (let i = 0; i < faker.number.int({
 			min: 1,
 			max: 50,
 		}); i++) {
-			records.push({
-				username: faker.internet.userName(),
-				date: faker.date.anytime().getTime(),
-				items: getChanges(),
-			});
+			const record = new TreeChangeModel();
+			record.username = faker.internet.userName();
+			record.date = faker.date.anytime().getTime();
+			record.items = getChanges();
+			records.push(record);
 		}
 		return records;
 	}
