+++ ui/src/models/TreeViewModel.ts
@@ -1,12 +1,13 @@
-import { DiffEntity, DifferApi } from "@incutonez/spec/dist";
+import { DifferApi, TreeItemModel } from "@incutonez/spec";
 import { configuration } from "@/apiConfig";
 import { ViewModel } from "@/models/ViewModel";
 
 export const DifferAPI = new DifferApi(configuration);
 
-export class DiffModel extends ViewModel implements DiffEntity {
-	field = "";
-	value = "";
+export class TreeViewModel extends ViewModel implements TreeViewModel {
+	username = "";
+	date = Date.now();
+	items: TreeItemModel[] = [];
 
 	static async readAll() {
 		const response = await DifferAPI.getDiff();
