+++ ui/src/types/components.ts
@@ -1,5 +1,12 @@
 // TODOJEF: Add Symbol for disabled options, and then set it up to use that in the optionDisabled fn
-import type { HTMLAttributes } from "vue";
+import type { HTMLAttributes, ObjectEmitsOptions } from "vue";
+// eslint-disable-next-line vue/prefer-import-from-vue
+import { UnionToIntersection } from "@vue/shared";
+
+// Taken from Vue source, as it's not exported by them...
+export type EmitFn<Options = ObjectEmitsOptions, Event extends keyof Options = keyof Options> = Options extends Array<infer V> ? (event: V, ...args: any[]) => void : {} extends Options ? (event: string, ...args: any[]) => void : UnionToIntersection<{
+	[key in Event]: Options[key] extends (...args: infer Args) => any ? (event: key, ...args: Args) => void : Options[key] extends any[] ? (event: key, ...args: Options[key]) => void : (event: key, ...args: any[]) => void;
+}[Event]>;
 
 export interface IOption {
 	id?: string | number;
